website/requirements.txt
altgraph==0.17.4
asgiref==3.4.1
asyncssh==2.17.0
attrs==21.4.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
cached-property==1.5.2
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2021.10.8
cffi==1.17.1
charset-normalizer==2.0.10
colorama==0.4.6
cryptography==42.0.8
cssselect2==0.7.0
defusedxml==0.7.1
Django==3.2.11
django-admin-interface==0.18.4
django-colorfield==0.6.3
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-jalali-date==1.0.1
django-optimized-image==0.3.0
django-tinymce==3.4.0
et-xmlfile==1.1.0
h11==0.14.0
html5lib==1.1
idna==3.3
isodate==0.6.1
jdatetime==3.8.1
lxml==5.3.0
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
paramiko==3.5.0
pillow==10.4.0
platformdirs==2.4.1
psycopg2-binary==2.9.10
pycparser==2.22
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyNaCl==1.5.0
pyphen==0.16.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2021.3
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
selenium==4.25.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.4.2
tinify==1.5.2
tinycss2==1.3.0
tqdm==4.66.5
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.8
WeasyPrint==52.5
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
xlrd==2.0.1
zeep==4.1.0
setuptools
uwsgi
# for ERD
django-extensions
pydotplus
graphviz

djangorestframework
django-filter

drf-spectacular